+++ src/test/suite/profile.test.ts
@@ -16,11 +16,11 @@ function test_profile_time(base: string, elf: string) {
 	fs.writeFileSync(path.join(testOutDir, base + '.time.amigaprofile'), profiler.profileTime(profileFile));
 }
 
-function test_profile_size(elf: string) {
+function test_profile_size(base: string, elf: string) {
 	const symbolTable = new SymbolTable(path.join(binDir, 'm68k-amiga-elf-objdump.exe'), path.join(testDataDir, elf));
 	const sourceMap = new SourceMap(path.join(binDir, 'm68k-amiga-elf-addr2line.exe'), path.join(testDataDir, elf), symbolTable);
 	const profiler = new Profiler(sourceMap, symbolTable);
-	fs.writeFileSync(path.join(testOutDir, elf + '.size.amigaprofile'), profiler.profileSize(path.join(binDir, 'm68k-amiga-elf-objdump.exe'), path.join(testDataDir, elf)));
+	fs.writeFileSync(path.join(testOutDir, base + '.size.amigaprofile'), profiler.profileSize(path.join(binDir, 'm68k-amiga-elf-objdump.exe'), path.join(testDataDir, elf)));
 }
 
 function test_unwind(elf: string) {
@@ -42,7 +42,10 @@ suite("Profiler", () => {
 		test_profile_time('amiga-profile-1590418304029', 'test2.elf');
 	});
 	test("Size: test3.elf", () => {
-		test_profile_size('test3.elf');
+		test_profile_size('test3', 'test3.elf');
+	});
+	test("Size: bitshmup.elf", () => {
+		test_profile_size('bitshmup', 'private/bitshmup.elf');
 	});
 	test("bitshmup.elf", () => {
 		//test_profile('amiga-profile-1589891749803', 'private/bitshmup.elf');